Replace `ts-loader` with `@babel/preset-typescript` and `fork-ts-checker-webpack-plugin` (#35713)
* Replace `ts-loader` with `@babel/preset-typescript` and `fork-ts-checker-webpack-plugin`

We've decided (pdWQjU-FO-p2) to prefer `@babel/preset-typescript` over
`ts-loader` for dealing with TypeScript in the context of Webpack. For
generating the definition files, we're generally going to use
`fork-ts-checker-webpack-plugin` (although `tsc --emitDeclarationOnly`
is also an option).

Also, as cleanup, these packages are switched to `moduleResolution:nodenext`
to ensure the `.d.ts` files are usable by consumers using that rather
than `bundler`.

Of the packages affected by this:

* image-guide apparently just had a dep on ts-loader but doesn't use it.
* videopress-core has no changes in the build artifacts.
* boost-score-api, react-data-sync-client, and svelte-data-sync-client
  all wind up with smaller bundles (by 14K–19K) now that Babel decides
  what needs transpiling based on the browserslist config, so they use
  native classes, generators, promises, and array-spread operators now.

This PR also adds a linting check to warn against re-introduction of
ts-loader, and mentions this in the relevant doc.

* Fix missing dependency

* changelog

* Revert 7aab051 and f94d463

---------

Co-authored-by: Mark George <thingalon@gmail.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/7958217937